refactor : Camera, Album 접근하는 코드 protocol로 분리
