+++ ReceiptManager/ReceiptManager/Source/View/ComposeScene/ComposeViewController.swift
@@ -187,7 +187,7 @@ final class ComposeViewController: UIViewController, ViewModelBindable {
                 let count = owner.viewModel?.receiptDataRelay.value.count ?? .zero
                 
                 if indexPath.row == .zero && count < 6 {
-                    owner.uploadImageCell(true)
+                    owner.showAccessAlbumAlert(true)
                 }
             }
             .disposed(by: rx.disposeBag)
@@ -254,6 +254,7 @@ extension ComposeViewController: SelectPickerDelegate {
     }
 }
 
+// MARK: - PHPhotoLibraryChangeObserver
 extension ComposeViewController: PHPhotoLibraryChangeObserver {
     func photoLibraryDidChange(_ changeInstance: PHChange) {
         getCanAccessImages()
@@ -288,16 +289,8 @@ extension ComposeViewController: PHPhotoLibraryChangeObserver {
 }
 
 // MARK: - UploadImageCell and Confirm Auth
-extension ComposeViewController {
-    private func requestPHPhotoLibraryAuthorization(completion: @escaping (PHAuthorizationStatus) -> Void) {
-        PHPhotoLibrary.requestAuthorization(for: .readWrite, handler: { authorizationStatus in
-            DispatchQueue.main.async {
-                completion(authorizationStatus)
-            }
-        })
-    }
-    
-    private func openAlbum() {
+extension ComposeViewController: CameraAlbumAccessAlertPresentable {
+    func openAlbum() {
         requestPHPhotoLibraryAuthorization { authorizationStatus in
             switch authorizationStatus {
             case .authorized:
@@ -320,15 +313,7 @@ extension ComposeViewController {
         }
     }
     
-    private func requestCameraAuthorization(completion: @escaping (Bool) -> Void) {
-        AVCaptureDevice.requestAccess(for: .video) { isAuth in
-            DispatchQueue.main.async {
-                completion(isAuth)
-            }
-        }
-    }
-    
-    private func openCamera() {
+    func openCamera() {
         requestCameraAuthorization { isAuth in
             if !isAuth {
                 self.showPermissionAlert(text: ConstantText.camera)
@@ -339,46 +324,6 @@ extension ComposeViewController {
                 self.present(picker, animated: true, completion: nil)
             }
         }
-    }
-    
-    private func uploadImageCell(_ isShowPicker: Bool) {
-        let alert = UIAlertController(
-            title: "영수증 사진선택",
-            message: "업로드할 영수증을 선택해주세요.",
-            preferredStyle: .actionSheet
-        )
-        
-        let cancelAction = UIAlertAction(title: "취소", style: .cancel)
-        let cameraAction = UIAlertAction(title: "촬영", style: .default) { _ in
-            self.openCamera()
-        }
-        
-        let albumAction = UIAlertAction(title: "앨범", style: .default) { _ in
-            self.openAlbum()
-        }
-        
-        [cameraAction, albumAction, cancelAction].forEach(alert.addAction(_:))
-        present(alert, animated: true, completion: nil)
-    }
-    
-    private func showPermissionAlert(text: String) {
-        let alertController = UIAlertController(
-            title: "\(text) 접근 권한 필요",
-            message: "\(text)에 접근하여 사진을 사용할 수 있도록 허용해주세요.",
-            preferredStyle: .alert
-        )
-        let settingsAction = UIAlertAction(title: "설정으로 이동", style: .default) { _ in
-            if let settingsUrl = URL(string: UIApplication.openSettingsURLString) {
-                UIApplication.shared.open(settingsUrl, options: [:], completionHandler: nil)
-            }
-        }
-        
-        let cancelAction = UIAlertAction(title: "취소", style: .cancel, handler: nil)
-        
-        alertController.addAction(settingsAction)
-        alertController.addAction(cancelAction)
-        
-        present(alertController, animated: true, completion: nil)
     }
 }
